+++ js/modules/appSystem.js
@@ -2,11 +2,13 @@ import { AppRegistry } from './taskbar.js';
 import SampleApp from '../apps/sampleApp.js';
 import VSCode from '../apps/vscode.js';
 import AppStore from '../apps/appStore.js';
+import ExportImport from '../apps/export.js';
 
 const AppSystem = (() => {
     const appModules = {
         sampleApp: SampleApp,
-        vscode: VSCode
+        vscode: VSCode,
+        export: ExportImport
     };
 
     function init() {